+++ devops/terraform/11_elasticache.tf
@@ -1,0 +1,16 @@
+resource "aws_elasticache_subnet_group" "production" {
+  name       = "main"
+  subnet_ids = [aws_subnet.private-subnet-1.id, aws_subnet.private-subnet-2.id]
+}
+
+resource "aws_elasticache_cluster" "production" {
+  cluster_id           = var.elasticache_name
+  engine               = "redis"
+  node_type            = var.elasticache_instance_class
+  num_cache_nodes      = 1
+  parameter_group_name = "default.redis6.x"
+  engine_version       = "6.2"
+  port                 = 6379
+  security_group_ids   = [aws_security_group.elasticache.id]
+  subnet_group_name    = aws_elasticache_subnet_group.production.name
+}