+++ devops/terraform/12_s3.tf
@@ -1,0 +1,38 @@
+resource "aws_s3_bucket" "frontend" {
+  bucket = var.s3_bucket_name
+}
+
+resource "aws_s3_bucket_policy" "frontend" {
+  bucket = aws_s3_bucket.frontend.id
+  policy = <<POLICY
+{
+    "Version": "2012-10-17",
+    "Id": "Policy1569048588734",
+    "Statement": [
+        {
+            "Sid": "Stmt1569048587703",
+            "Effect": "Allow",
+            "Principal": "*",
+            "Action": "s3:*",
+            "Resource": "arn:aws:s3:::${var.s3_bucket_name}/*"
+        }
+    ]
+}
+POLICY
+}
+
+resource "aws_s3_bucket_public_access_block" "frontend" {
+  bucket = aws_s3_bucket.frontend.id
+}
+
+resource "aws_s3_bucket_website_configuration" "frontend" {
+  bucket = aws_s3_bucket.frontend.bucket
+
+  index_document {
+    suffix = "index.html"
+  }
+
+  error_document {
+    key = "error.html"
+  }
+}