+++ minime/views.py
@@ -1,7 +1,9 @@
-from django.http import HttpResponseRedirect
+from django.conf import settings
+from django.http import HttpResponse, HttpResponseRedirect
+from django.views import generic
 from django.shortcuts import get_object_or_404, render
 from rest_framework import generics
-from minime import forms, models, serializers, utils
+from minime import models, serializers, utils
 
 
 def index(request):
@@ -11,7 +13,11 @@ def index(request):
     :param request: Request object
     :return: Template http response
     """
-    return render(request, "index.html", {"form": forms.ShortenURLForm})
+    context = {"debug": settings.DEBUG}
+    if not settings.DEBUG:
+        context["react_urls"] = utils.build_react_urls(utils.get_current_app_version())
+
+    return render(request, "index.html", context)
 
 
 def redirect(request, hash):
@@ -26,10 +32,8 @@ def redirect(request, hash):
     url = utils.cache_get_short_url(hash)
 
     if not url:
-        urlobj = get_object_or_404(
-            models.Url, hash=hash
-        )  # gets object, if not found returns 404 error
-        url = urlobj.url
+        urlobj = get_object_or_404(models.Url, hash=hash)
+        url = getattr(urlobj, "url")
         utils.cache_set_url(hash, url)
 
     utils.create_stats(request, hash)
@@ -46,3 +50,13 @@ class CreateShortUrl(generics.CreateAPIView):
 
     queryset = models.Url.objects.all()
     serializer_class = serializers.UrlSerializer
+
+
+def ping(request):
+    return HttpResponse(status=200)
+
+
+class ReactResetView(generic.View):
+    def get(self, request, *args, **kwargs):
+        utils.get_current_app_version(use_cache=False)
+        return HttpResponse("reset")